Puedo calcular distancias!!

+++ src/main/java/airlinetycoon/modelo/Posicion.java
@@ -2,21 +2,21 @@ package airlinetycoon.modelo;
 
 public class Posicion
 {
-	private float latitud;
-	private float longitud;
-	public float getLatitud()
+	private double latitud;
+	private double longitud;
+	public double getLatitud()
 	{
 		return latitud;
 	}
-	public void setLatitud(float latitud)
+	public void setLatitud(double latitud)
 	{
 		this.latitud = latitud;
 	}
-	public float getLongitud()
+	public double getLongitud()
 	{
 		return longitud;
 	}
-	public void setLongitud(float longitud)
+	public void setLongitud(double longitud)
 	{
 		this.longitud = longitud;
 	}
